+++ trunk/src/com/geoscope/GeoEye/Space/TypesSystem/CoComponent/ObjectModel/GeoMonitoredObject1/DEVICE/ControlsModule/Model/Data/ControlStream/TDataStreamPanel.java
@@ -46,7 +46,7 @@ public class TDataStreamPanel extends Activity {
 	private String 	ServerAddress;
 	private int 	ServerPort;
 	//.
-	private int 	UserID;
+	private long 	UserID;
 	private String 	UserPassword;
 	//.
 	private TCoGeoMonitorObject 	Object;
@@ -75,7 +75,7 @@ public class TDataStreamPanel extends Activity {
             	ServerAddress = extras.getString("ServerAddress");
             	ServerPort = extras.getInt("ServerPort");
             	//.
-            	UserID = extras.getInt("UserID");
+            	UserID = extras.getLong("UserID");
             	UserPassword = extras.getString("UserPassword");
             	//.
         		TReflector Reflector = TReflector.GetReflector();
