+++ trunk/src/com/geoscope/GeoEye/Space/TypesSystem/Visualizations/TileImagery/TTileImagery.java
@@ -177,7 +177,7 @@ public class TTileImagery {
 	private void HttpServer_LoadDataFromServer() throws Exception {
 		String URL1 = Reflector.Server.Address;
 		//. add command path
-		URL1 = "http://"+URL1+"/"+"Space"+"/"+"2"/*URLProtocolVersion*/+"/"+Integer.toString(Reflector.User.UserID);
+		URL1 = "http://"+URL1+"/"+"Space"+"/"+"2"/*URLProtocolVersion*/+"/"+Long.toString(Reflector.User.UserID);
 		String URL2 = "TileServerData.dat";
 		//. add command parameters
 		URL2 = URL2+"?"+"0"/*command version*/;
